Update patch set 3
Patch Set 3:

Indeed once a redirect is created the old revisions can then be restored.

I don't think a maint script would make sense as this is a case that can keep occurring.

I go and create an item for enwiki:FooBar, the item is deleted and the sitelink manually moved to another item.
I then ened to restore the first item and create a redirect!

Patch-set: 3
Label: Verified=0