Update patch set 2
Patch Set 2: Code-Review-1

(2 comments)

Mistake in the documentation.

This new interface is much nicer, thanks. I still think other usages of PropertyInfoStore should be changed to more restricted interfaces. Currently PropertyInfoStore::getPropertyInfo is exclusively used in PropertyInfoDataTypeLookup::getDataTypeIdForProperty. But this can not be done in this chain of patches, obviously.

Patch-set: 2
Label: Code-Review=-1