Make UsageTrackingIntegrationTest more robust and not leak data
That test created items and properties with hard coded ids,
but never deleted them. That made other tests fail in certain
cases. There's no need for that test to create entities at all,
as we do usage tracking regardless of the existence of the entity.

This also fixes Scribunto_LuaWikibaseLibraryTestCase which made this
test break for me sometimes (as Scribunto_LuaWikibaseLibraryTestCase
changed the usage tracker to a NullUsageTracker which obviously
the integration test didn't like).

Also made the test setting independent.

Change-Id: Icbe3563945a3731a2415249b8769b31e0eb7e18e