Update patch set 3
Patch Set 3:

(4 comments)

The specific reason for some of the changes are unclear to me.

In any case, we should have a single helper function for comparing sets of triples, and a single helper function for getting a normalized list of triples. If we are touching all these usages anyway, we might as well introduce them now, instead of making the same changes in all locations.

Patch-set: 3