Remove duplicate PropertyDataTypeMatcher from lib
Note the difference between these two implementations. The one in lib
did caching, while the caching aspect is split of in DataModelServices.
I'm re-adding the caching aspect *only* for the code that used it before.
Introducing this caching level for other usages of the same lookup should
be done in an other patch.

Change-Id: I51a8cf19375122dd82921a73f9f94b850ad3a4b6