Update patch set 3
Patch Set 3: Code-Review-1

(3 comments)

@thiemo: thanks for Icb433d502, that is going into the right direction.

As for this change: I oppose dropping the normalization code. I rather suggest we do more normalization, and more consistently.

The RDF tests are not about the output syntax at all. They test the RDF mapping. The output syntax is irrelevant to that. We use n-triples for convenience, and any equivalent set of n-triples should match.

If we want to test rdf serialization, that should be done in a test for the rdf writers, not the rdf mapping.

Patch-set: 3
Reviewer: Daniel Kinzler <128@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1