Update patch set 3
Patch Set 3:

I'm a bit town. On the one hand, if only the code is needed, we should pass only the code, to reduce our dependencies to a minimum. On the other hand, the code is a plain string, so it's not type safe, and not future proof (what if we happen to need something more than just the code somewhere deep down at one point?).

The best solution would probably be a LanguageId class, but introducing that into core is a whole different can of worms.

So... no strong feelings about this. This patch is an improvement, even though I'm not entirely happy about it, especially with regards to performance:

Instantiating Language objects is VERY expensive. As long as we only use the one or two that have been loaded anyway, fine. But we should be careful not to load dozens or even hundreds. Please double-check the implications here.

Patch-set: 3
Reviewer: Daniel Kinzler <128@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0