Update patch set 3
Patch Set 3:

(3 comments)

> I oppose dropping the normalization code.

I'm sorry? I'm not "dropping the normalization code". Where does this come from? All normalization that's necessary to make the tests pass is still there.

> The RDF tests are not about the output syntax at all.

So? This is not a reason to hide obvious errors in the test setup and/or the code. You can not argue that the output format is irrelevant. The whole test setup is build upon a specific output format. Try to change it and you will see how "irrelevant" it is. The format is parsed and used for the final assertions. The parser used for that must be as simple as possible. It must only do what's necessary to make the tests pass. Nothing more. Otherwise you would need to extract that full featured parser into it's own class to make it testable. I'm still not sure this is worth the trouble. As you can see, the tests work just fine with a much more minimal, super-trivial parser step that does not involve radical trimming all over the place.

Patch-set: 3