Update patch set 3
Patch Set 3:

also, wonder if we deploy the code to the repo a day before all the clients, then think we can also have this situation.

i suppose we can wait until the code with the new data type is available everywhere before converting any properties.

maybe that is okay enough and any new external identifier properties might not get used that quickly in the clients...

Patch-set: 3