Enable CyclomaticComplexity and NestingLevel PHPCS sniffs
The numbers reflect the current "worst cases" in our code base. The effect
of this is: Whenever a commit introduces code that's more complex than
what's currently in the code base, we will be aware of that. In most cases
we will simply raise the limit in the same commit. So this is not about
blocking such commits. It should be allowed to raise these limits if it
makes sense. This is about awareness.

This excludes two cases with a CyclomaticComplexity of 22 and 21. I tried
to refactor these two methods, but all other ways of expressing the same
are way harder to read and to maintain. These two switch statements are
*not* bad code.

Also see:
https://github.com/wmde/WikibaseDataModel/pull/589
https://github.com/wmde/WikibaseDataModelServices/pull/101

Change-Id: Icb53bda3ef3307e0e08742a3f966092db829bf2d