Update patch set 15
Patch Set 15: Verified-1 Code-Review-2

(5 comments)

This still fails for the same reason as already noted in several comments. I think I even gave a finished solution: simply copy and paste the line I wrote and be done. Why was this not done? This also still contains unrelated changes I asked to revert several patch sets ago. Do the comments get lost somehow? I know Gerrit is not nice to follow. But I really do not know how I can make the issues with this more obvious.

So the answer is no. Not even the PHPUnit test responsible for the class touched in this patch does succeed. I would argue this is the most basic condition to call this exercise "done".

On a wider scale and unrelated to the exercise: I'm not sure if we can ever merge this as it is. It's a breaking change. People relying on this API will complain. Should the default be backwards compatible, including the URL? Probably not. So we need to advertise this change. We need to check who relies on this and tell them before we make the breaking change. I'm setting a temporary -2 for this reason.

Patch-set: 15
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=-1
Label: Code-Review=-2