Resolve language confusion in terms view
There is quite some confusion going on here.

Note that this class does have 3 public methods.

When requesting HTML via getHtml() (this is for the label,
description and aliases on top of the page) and getTitleHtml() the
$this->languageCode is *not* the interface language but the language
of the requested term. Messages use wfMessage(), which defaults to
$wgLang, which is set to qqx in the test.

But when calling getEntityTermsForLanguageListView() the same
$this->languageCode property is misused as the interface language.

Luckily this is not an issue in production because
$this->languageCode is always set to the users interface language.

This also splits a test with a bad data provider (note the removed
$numberOfPlaceholders and substr_count) and adds a bunch of missing
tests and assertions.

Change-Id: I4270a15e8f3e516d154d1a4d1018a9760c5ba95d