Correctly exclude disabled PHPCS rules
With the <severity>0</severity> hack the specific sniff is still run,
but not reported any more. It still consumes CPU cycles but does not
do anything.

This is not critical in these two cases, because this only disables
specific messages (it's "a.b.c.d") and not rules (that would be
"a.b.c").

Change-Id: I9fd26d4ba78efa4231374a0d6c26baf91615d674