Correctly exclude disabled PHPCS rules
With <severity>0</severity> the sniff is still run, but the disabled
error is not reported any more. This is not critical in cases where
only one of many possible messages is disabled.

But this is not the case here. The sniffs touched in this patch have
only one message. Disabling this message leaves nothing to report.
The sniff still runs and consumes CPU cycles but never reports anything.

Bug: T121291
Change-Id: I9fd26d4ba78efa4231374a0d6c26baf91615d674