Temporary disable PHPCS sniffs for performance reasons
I did an actual profiling. These 3 sniffs are responsible for 15%
of the total runtime. Not a "golden bullet", unfortunatelly. It's
simply the total number of files in our repository, multiplied with
the total number of sniffs.

However, I suggest to merge this for now, as long as the timeout
issue is not fixed. This unblocks the current situation where
patches constantly get stuck. Obviously, we will revert this patch
later.

Bug: T121291
Change-Id: I73209463d23099f23a07d64e07bbd0ecff69cb2d