Delete at least one second worth of data in ChangePruner
It's not too bad to go slightly over the batch size
here as we're just going over it by one second worth
of data (which is never going to be more than a thousand
rows). A deletion of ~1500 rows (as the default batch
size is 500) is not nice, but also not worrisome.

Bug: T122336
Change-Id: I37ced5f3fb3ed4a16632980b47f6f588c0839bef