Inline Message::params() when not needed
In patch set 2 I reworked this patch completely. I do not swap "wfMessage()"
with "new Message()" or the other way around. Instead I'm focusing on
changes that make the code easier to read, easier to understand and easier
to maintain.

I'm mostly removing/inlining Message::params() calls when they do not do
anything special. This method is useful when you want to *append*
parameters, but this is not the case in most places.

I'm also fixing a bug where a Message object is passed to a function that
requires a string, and I'm spliting some long lines.

Change-Id: Ia54b9f710ca0dc5e5e1183282cf2c2970918505a