Replace Iterator with Traversable in method parameters
This is split from I5a2e356. This patch focuses exclusively on @param
tags and does not touch any @return tag.

For all methods that are private I narrowed the interface as much as
possible. In most cases the object passed in is a ResultWrapper object,
but never an array. But neither this nor Iterator are the relevant
interfaces for methods that do nothing but iterating via foreach. The
only relevant interface for this is Traversable.

Change-Id: I15f16da3cedac9b0d9ea7b83cde1914d2ef60d0d