Update patch set 3
Patch Set 3:

I reworked the thing completely. I also got rid of the hack in TruthyStatementRdfBuilderTest where a property was glued on an object on runtime. Now I find this tests code way, way easier to read and understand.

I reverted my attempt to extract the getTestData method (this code is repeated 6 times in 6 tests). I will retry this in a later patch.

Patch-set: 3
Label: Verified=0