Rework RDF builder tests to get rid of object property hack
This is the most minimal patch I can think of to get rid of the hack
(a property is glued onto each $builder object on runtime). I wanted
this patch to focus exclusively on this single detail. There is more to
do and I will continue working on this in later patches.

Change-Id: I01759e8efd0ddcc7431ba7c25b063e2b6d3063ce