Fix tests failing because of HTMLForm change in core
This was changed in I7e15331.

Note: We are aware that we need to change these tests. We should not
use the deprecated assertTag any more, and these tests should not test
the <button> tag. But this is the most simple fix to unbreak our CI.
We will continue working on this.

Bug: T121584
Change-Id: If2b68e0f9e916ec10678c7e1a11099acea1a9cce