Update patch set 2
Patch Set 2:

To understand what's going on in this patch, I need to understand what a "label revision" is, and what the difference to the "base revision" is. My question is: What's wrong with the old code?

In general, tracking the revision is done for the same reason as in every WikiEditor: Resolving conflicts. This must be done in the clients JavaScript, because what you edit is based on a specific revision. Who else should keep track of this, if not the client doing the edit?

If you don't know the revision an edit is based on, all you can do server-side is to accept everything, which will silently overwrite 50% of concurrent edits.

Patch-set: 2
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0