Increase EntityHandler code coverage
This class is reported as one of the top project risks (currently on rank #3
after EditEntity and UpdateRepoHookHandlers) in PHPUnit's coverage report.

This patch is only a small step. The main issue is that the large (and hard
to test) methods getEntityDeletionUpdates and getEntityModificationUpdates
are not covered by any test. I will continue working on this in a later
patch.

Change-Id: Ib357080095f6a42659ac9acd78b4ee986b2f1fe3