Do not fail hard on HistoryAction because of an undeserializable revision
This fixes a very specific issue: When the newest revision of a page
contains a JSON blob that can not be deserialized, the whole action fails
hard with an exception. But why? There is only one minor detail where the
deserialized entity (or redirect) is used: for the title on top of the
history page. If this fails, we can simply fall back to the title of the
wiki page.

This came up while working on
https://github.com/wmde/WikibaseDataModel/pull/593 (Do not allow
self-redirects).

This does *NOT* fix the issue that such a page with the newest revision
being undeserializable can not be restored. This needs more investigation
and a more complicated fix.

Change-Id: Ie07d170053140daa8022431f451e0761c33f55af