Make use of EntityId::__toString for readability
I believe this increases the readability of the code. EntityId does have
a well defined __toString method. Technically, __toString can change because
it's made for debugging and messages presented to the user. That's exactly
what I'm using it for in this patch. This and in test setups/mocks/dummies.

Practically, Entity::__toString is extremely unlikely to change. There is
nothing else in this object. But even with such a guarantee, __toString
should not be used in production code to form identifiers and such.

Change-Id: I8d309410109e2e3149e6d76f1b760eaa06190cb7