Update patch set 4
Patch Set 4:

Hoo, I will continue working on this in later patches. It would be nice to have this patch merged as it is, if it is ok. I do not want to make it more complicated.

Patch-set: 4
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0