Resuse NTriplesRdfTestHelper in as many tests as possible
That's all I could find.

I want to continue working on this. For example, I want to play around
with the RdfBuilderTestData class. It looks like it could be merged
with the helper I introduced. Or the helper could be a member of the
RdfBuilderTestData class. This should reduce code duplication even
more. But I want to do this in later patches.

Change-Id: Ifc68eebc6d62983f07bebdbe0496b2feba47ce6e