Add a "WikibaseClientOtherProjectsSidebar" hook
In order to be able to overwrite parts of the sidebar
for Wikimedia specific things.
For now we plan to use this in order to show links to
commons categories in the sidebar.

I figured it would be much safer to validate the data
we get from the hook handler and given the size of the
array it really isn't expensive. If you still think the
data shouldn't be validated for some reason, I'm ok
with removing that part.

Bug: T94989
Change-Id: Ica9b1bac5394cddc021c013da4e6b5a74651c4aa