Filter data type registrations with disabledDataTypes setting
Removed the old (unused) dataTypes setting from lib and
introduced a new 'disabledDataTypes' setting in client + repo.

The disabledDataTypes setting can be used to define which
data types are disabled for a wiki. Added documentation
for the setting.

The new data types registration system wasn't considering
the dataTypes setting. We now have a hook to allow other
extensions to add data types (e.g. math), so a setting
that lists enabled types doesn't work so nicely with that.
A setting that does the inverse can work for this.

At least for Wikidata, we sometimes need to set this explicitly
so we can deploy new data types in a more controlled way.

The ability to enable only certain types might also be
desired by some third party users.

Bug: T123447
Change-Id: I07abc6bb5a785e2331bdc98412cf45a7160c5b9a