Update patch set 4
Patch Set 4:

(1 comment)

Why is this merged with the issue not really fixed? In my opinion the code is  not better now, it's worse because it now requires an object of a way to generic type (IContextSource). The real type is now hidden in an instanceof check. Before it was visible in the documentation.

So why is this merged?

Patch-set: 4
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0