Require MutableContext as constructor param in EditFilterHookRunner
This is a direct follow-up to the comments in I3e82e8f.

Technically this patch does not do much. If the $context object provided
does not implement both interfaces, the code will fail. This is not
changed by this patch. But the place where the failure happens is
changed. Instead of an undocumented feature-detecting "if" hidden
somewhere in the guts of EditFilterHookRunner, the type is now visible
in the public interface. It's even a PHP type hint and not only a @param
tag.

Change-Id: I566cf24cb1ed7fe721a744acc0d2121d39d82a1e