Update patch set 1
Patch Set 1: Code-Review-1

I agree this is not too clear. From my point of view, $languageCode is always both: The language to show (if there is only one) and the language in which to show. The latter is not fully implemented where `wfMessage` is called directly. Also, it doesn't make too much sense binding language to show and language in which to show together. So I suggest to start with replacing `wfMessage` calls by `$this->msg` calls. A second step could be introducing a `$languageCode` parameter to `getHtml` and `getTitleHtml`.

Patch-set: 1
Label: Code-Review=-1