Update patch set 1
Patch Set 1:

I think I understood what you are trying to do. You are trying to consistently use `$this->languageCode` as the content language for `getHtml` and `getTitleHtml` and `$wgLang` as the UI language. While this would be more consistent, I think it's not desirable. For example, it adds an unnecessary usage of a MediaWiki global. 

Instead, I want `$this->languageCode` to consistently be the UI language and have the content language consistently as a method parameter.

Patch-set: 1