Update patch set 1
Patch Set 1:

There is only 1 parameter. Introducing a 2nd one is not part of this patch.

This is the terms view. The first question a user of this code asks is not what the UI language should be. The first parameter answers the question which term you want to see. That's what I'm using the parameter for. Nothing more, nothing less.

This  patch fixes an issue and make the code more consistent and reliable. I really do not understand why this deserves a -1.

This patch does not even introduce a dependency on core. It's already there. How can you argue this patch makes things worse, when it does not?

Introducing a 2nd parameter is not blocked by this patch. This can easily be done in a later patch.

Patch-set: 1
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0