Update patch set 6
Patch Set 6:

how do you implement the hook that only needs the title and not the id?

not so nice, but https://gist.github.com/filbertkm/2909cf059701d7a9b99d is what i came up with and it needs item id. (we need to make sure this takes care of escaping, etc.)

also, it would be nice if commons category could go into parser output and somehow get it from there if it is, and if not do lookup. (and eventually remove doing lookup, once cache epoch is bumped enough)

also, does this hook run on parse only? or page views? (i realize sometimes there is sidebar caching and of course varnish)

if it's only on parse, then a lot of pages will be missing commons category or whatever.  but it's not nice to do lookup of the commons category on page views.

Patch-set: 6
Label: Verified=0