Update patch set 6
Patch Set 6: Code-Review-1

The hook handler gets the item id, not the title. That's maybe a little hard to spot given I forgot to change the signature or runHook to take the id directly and not take the title and then get the id from the DB again.

Patch-set: 6
Label: Code-Review=-1