Update patch set 4
Patch Set 4:

https://www.mediawiki.org/wiki/Wikibase/DataModel#Geographic_locations says:

a precision (decimal, representing degrees of distance, defaults to 0, 9 digits after the dot and three before, unsigned, used to save the precision of the representation)

So default is 0 and null should not exist at all. Looks like converting null to default 0 is the correct way? Or the docs should be fixed?

Patch-set: 4
Label: Verified=0