Update patch set 1
Patch Set 1:

Dependency on $wgLang is new. I gave this change a -1 because I think it moves in a direction I don't want this code to go.

There is only one language code that all three public methods need: The UI language. This language code could be a constructor argument. It does not make sense to have a language code as constructor argument and then not use it in one of the public methods. I can imagine a few good interfaces here:

a) (What I proposed)
  __construct(uiLanguage)
  getEntityTermsForLanguageListView(contentLanguages)
  getTitleHtml(contentLanguage)
  getHtml(contentLanguage)

b) (both in construct)
  __construct(uiLanguage, contentLanguages)
  getEntityTermsForLanguageListView()
  getTitleHtml() uses contentLanguages[0]
  getHtml() uses contentLanguages[0]

Barely ok:

c) (contentLanguages in construct)
  __construct(contentLanguages)
  getEntityTermsForLanguageListView(uiLanguage)
  getTitleHtml(uiLanguage) uses contentLanguages[0]
  getHtml(uiLanguage) uses contentLanguages[0]

d) (Always explicit)
  __construct()
  getEntityTermsForLanguageListView(uiLanguage, contentLanguages)
  getTitleHtml(uiLanguage, contentLanguage)
  getHtml(uiLanguage, contentLanguage)

e) (contentLanguage in construct but used in getEntityTermsForLanguageListView)
  __construct(uiLanguage, contentLanguage)
  getEntityTermsForLanguageListView(contentLanguages[1..]) uses contentLanguage as contentLanguages[0]
  getTitleHtml()
  getHtml()

f) (contentLanguage in construct but checked against in getEntityTermsForLanguageListView)
  __construct(uiLanguage, contentLanguage)
  getEntityTermsForLanguageListView(contentLanguages) asserts contentLanguage equals contentLanguages[0]
  getTitleHtml()
  getHtml()

What you are doing in this patch:

  __construct(contentLanguage)
  getEntityTermsForLanguageListView(contentLanguages) does not do anything with contentLanguage
  getTitleHtml()
  getHtml()

I see two issues here:
  - uiLanguage is not configurable anymore for getEntityTermsForLanguageListView
  - The constructor argument and the argument to getEntityTermsForLanguageListView are not put in context with each other

For completeness, what we currently have:

  __construct(someLanguage)
  getEntityTermsForLanguageListView(contentLanguages) uses someLanguage as uiLanguage
  getTitleHtml() uses someLanguage as contentLanguage and takes uiLanguage from global scope
  getHtml() uses someLanguage as contentLanguage and takes uiLanguage from global scope

So, just to be clear: Your patch makes stuff better since it makes stuff more consistent. Doing it in a way I would actually like would still mean reverting most of this change, so I don't want to have it merged in the current state.

Patch-set: 1