Update patch set 3
Patch Set 3:

I would prefer to call this "wikibase-propertytypes". Even though the property types are used as keys in the array, and the value types as values, the idea is to provide information about propertytypes.

Also, using objects as values in the array, instead of simple strings, would allow additional information about property types to be included in the future, without breaking clients.

Patch-set: 3