Update patch set 8
Patch Set 8: -Code-Review

I'm removing my vote for now, mostly because this change is not relevant in production. We do have a validator in place that does not allow null-precision on coordinate values. The precision is auto-detected on parse, and the detected precision is stored. Null is never stored.

That's also the main reason why the line ...

 $precision = 1 / 3600;

... in GeoCoordinateFormatter is *not* relevant. This line is never executed in production. Instead the precision detection code in GlobeCoordinateParser is relevant, and this is much more complicated.

However, I think a fallback with 1/3600 is fine for the moment. It's better than whats currently in this code, and probably more useful than 0 to the user.

Patch-set: 8
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=0