Update patch set 8
Patch Set 8: Code-Review-1

found an issue, relating to implementing the commons category stuff.

if a connected item has a commons sitelink (the gallery), then "Wikimedia Commons" (with or without the hook) is the first item in the sidebar.

if the connected item has *no* commons sitelink, then "Wikimedia Commons" is the last item in the sidebar.

this inconsistency is probably going to annoy some people.

maybe OtherProjectsSidebarGenerator::sortAndFlattenSidebar could instead happen sometime after the hook is run.  This way, the sorting can be more consistent and the hook can also have access to the array keys (with site id) instead of more hacky $entry['msg'] === 'wikibase-otherprojects-commons' check there.

Patch-set: 8
Label: Code-Review=-1