Update patch set 2
Patch Set 2:

(1 comment)

What's unclear about what I wrote in the commit message? It seems you have not understood it. Does it help when I try again to explain?

Before this patch, the EditFilterHookRunner class was documented to require an IContextSource. But this was not even close to what really happend. In reality, it required a MutableContext. But this information was hidden in an if. Not documented anywhere. The code was lying.

After this patch, all this uglyness does not change much. The fact that an IContextSource object "magically" turns into an MutableContext object, but can make the code fail hard with an exception if this is not possible, all this is still there. But it's way more visible. It's documented in multiple interfaces now. And not only in "soft" @param type hints but in "hard" PHP function parameter types.

This is the whole intention of this patch. It makes uglyness more visible. You can not downvote me because of this. I neither introduced it nor did I wanted to fix it in this patch.

Feel free to add a patch on top of this or Iaf2a310 if you have an idea.

So what's the way forward here?

Patch-set: 2