Update patch set 2
Patch Set 2:

In that case someone will need to check all callers to this method and make sure we always get MutableContext objects.

I'm totally in favor of making problems obvious, but site reliability is more important. I'm willing to dig deeper into this in the next couple of days or maybe over the weekend. But I find this to dangerous to deploy :/

Patch-set: 2