Consistent use of white background in term box
This removes a misplaced white background from the aliasview, and adds
a white background for the whole terms table. This is for consistency,
because statements also have a white background. You have to change
the background color of the page to see the effect of these changes.
Otherwise they are "invisible" (white on white).

Change-Id: Ie25de234026412dcb8b32c10d267b0daed23e718