Update patch set 1
Patch Set 1:

Rank of a statement set to preferred within one statement group has no bearing on its rank within all statements of a given item. It should therefore not overwrite the image we got from a image property higher in the list of properties to use for the page image.

Patch-set: 1