Use Title::getPrefixedText where appropriate
Instead of Title::getFullText which also includes the section
fragment.
We already do that in many places, but there are also quite some
where we use the wrong value (it's wrong as we only handle full
titles, Wikibase doesn't know about sections).

As suggested by Katie in T47839#1954505.
I don't think this will help with the bug… but it's a good thing
to do in general.

Change-Id: I2e6803210a0ed0bbf16b036bfd739e8f72361a0c