Update patch set 1
Patch Set 1:

How is this "worse"? Can you explain this more?

Loops are cheap. It's not the number of loops that create a performance issue, it's what the loops *do*. With this patch *and* with what I have done in DataModel JavaScript 2.0 I belief this is faster or at least equal.

Patch-set: 1
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0