Update patch set 1
Patch Set 1:

Thanks for the patch. Personally, I'm not sure about this because it scratches a conceptual issue: The Wikibase software is a general thing and should not rely on things that exist on wikidata.org only. I will raise this as a question in the team.

Patch-set: 1
Label: Verified=0