Remove bad language codes from mono languages
In JS, we use a list of 391 language code (`mwfile`). In PHP, we have 416
language codes (`mw`). That's why we have to remove more language codes from
the PHP instance.

Also, since `wikibase.WikibaseContentLanguages` is not yet used for terms, I
hard-code the removal in that class instead of creating a JS implementation of
`DifferenceContentLanguages`. I would like to defer that since we probably want
to split out `LanguageNameLookup` beforehands as we did in PHP.

Bug: T125063
Change-Id: I8b203aae1c45efceb4459b62b166bab3c191afcd