Update patch set 1
Patch Set 1:

I assume the duplication you are referring to is in the Parameters method?
I could possibly write a followup to this chain to try to reduce the duplication here, I guess it would essentiall be an object which would provid different types of params for this method?

Patch-set: 1
Reviewer: Addshore <106@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0