Update patch set 1
Patch Set 1:

Adam, the events come in (at least this is how they come in in my local environment) in the following order:
* mwLoadStart
* mwLoadEnd
* wbInitStart
* wbInitEnd

I suggest to set up two graphs: One for wbInitEnd - wbInitStart and one for the total wbInitEnd - mwLoadStart.

Adrian, I'm not absolutely sure if these are the best positions for these lines. Please check.

Patch-set: 1
Label: Verified=0