Update patch set 1
Patch Set 1:

Most of this code is comments, and then the properties of each api class.
But it finally kills something we want wanted to get rid of for years.
Once this is merged we can continue slowly refactoring these api things as and when we need to!

Patch-set: 1
Label: Verified=0