Rename SnakConstructionService to SnakFactory
This now makes sense after the original class called "SnakFactory"
was removed in Ibda8131.

The difference between the two is: The old required a DataValue object.
The new uses a DatavalueFactory to construct this from a raw value.
It still is a SnakFactory, in my opinion.

One thing we could do in addition is to rename "newSnak" to
"newFromRawValue" (vs. "newFromDataValue"). Should I do this in this
patch?

Change-Id: Ie3956931e47745bbcd3c90cb9ac753059f9610e0