Helpful exception message when Wikibase Client/Repo not set up properly
This is relevant when you are adding (for example) the ArticlePlaceholder
extension to your LocalSettings.php, but you forgot to add
WikibaseClient.php too. Code in article placeholder will call
WikibaseClient::getDefaultInstance(), but this will fail with totally
non-helpful exceptions, because the default config is not loaded.

If you think there is a better way to help the user fix the problem,
please tell me.

Change-Id: I6205b853afa32c29003d36e00e026b898439da3a