Count calls to WikiPageUpdater methods
This will give us an accurate reference for
the number of updates being triggered by wikibase.

Vaugly relates to:
https://phabricator.wikimedia.org/T124418

Of course we may also not want to bother with this
and decide knowing how many changes we dispatch to
clients is enough.
Just a possibility..

Depending on how many times these methods are called
this may even be a bad idea and flood statsd.

Change-Id: Ie9a69822f7d8677e39b866437cebb8ce88603147