Update patch set 4
Patch Set 4: Code-Review+2

There already is a configuration change for this and it points the URL to the service description, that makes sense. 

I've amended the documentation to say that this should point to the service description.

Patch-set: 4
Reviewer: Hoo man <50@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+2