Update patch set 1
Patch Set 1:

I just wanted to say that this is probably not the correct fix. What this does is making class files in this directory available to be used in other files. This is not the case here.

Patch-set: 1
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0