Refactor ReferenceList
This removes the inheritance between ReferenceList and SplObjectStorage
and replaces it with a native array. To make the change as minimal as
possible, the Countable interface and the IteratorAggregate interface are
now implemented by ReferenceList so that it can be used almost the same as
before (like iterating over it and calling count() on it).