Add Math property type to ontology.owl
We discussed this and we think this is a sensible workaround. Technically
this is in the wrong code base. This belongs to the Math extension. But
how could the Math extension extend this documentation? This should be
good enough. Sure, this should be reconsidered when we start adding
dozens of property types.

Please tell us if you have a better idea.

Bug: T67397
Change-Id: I60b2259f261389e9ef81d3f0d9bc7a2d16974710