Update patch set 1
Patch Set 1:

For the record, I'm not too happy about having the info about an extension provided data type here. But we don't have a good alternative at the moment. I created T126355 to track that.

Note that datatype URIs don't have to resolve to anything. Providing an OWL description of the type is nice, but it's not a requirement. 

I'm tempted to give this CR-2. This is conceptually ugly, and there is no need to do this. However, if we use a custom URI later, that will be a breaking change to the data set. Clients and queries that rely on the old type URI will break.

In any case, this should probably not be merged until it is clear what format the literals in the datatypes have.

Patch-set: 1
Label: Verified=0