Make ChangeOp::validate run against the current revision.
The idea is that validation should happen against the current revision,
because modifications of the edit's base revision get patched into the
current revision later. If we perform validation against the base
revision, we may miss conflicts that arise only after patching.

This change also moves loadEntity() from SpecialWikibaseRepoPage to
SpecialModifyEntity, to allow access to the EntityRevisionLookup.
loadEntity() is not used outside of SpecialModifyEntit.

Bug: T121395
Change-Id: I0d49d676bc15efff4a96a3b06d3df26f3494c7a9