Update patch set 1
Patch Set 1:

I don't think in this case RDF export should contain anything but the actual data that the property contains. MathML AFAIK is a markup language, and as such I don't see how it is related to RDF. In any case, this is about documenting the type, so if we have type "Math" that RDF export produces we must document it in the ontology.

Patch-set: 1
Label: Code-Review=0