Update patch set 1
Patch Set 1:

I still do not understand what I should do here. As I already tried to explain: This is documenting the current status quo. No matter what you do to this file, this will not change anything in the RDF export.

Can somebody please either amend this patch to help me understand what can be improved here, or merge it as it is, and then continue the unrelated discussion on Phabricator? Gerrit is the wrong place for this.

Patch-set: 1
Label: Verified=0