Update patch set 1
Patch Set 1: Code-Review-1

So I have a local metawiki which doesn't have Wikibase client enabled (I control that by not including the entry point for that wiki). If I enable the ArticlePlaceholder there and visit Special:AboutTopic I get a fatal error because the  WikibaseClient class doesn't exist at all (it's not in the autoloader on my metawiki).

The case this patch would cover is when someone somehow includes the autoloader/ class map, but not the actual extension entry point. I'm not sure how that would happen, but it can't be ruled out. (Maybe by including the entry points but setting $wgEnableWikibaseClient/ $wgEnableWikibaseRepo to false?)

I think you should check against the "WB_VERSION" and "WBC_VERSION" constants, as these are actually supposed to be used for these kinds of detection (or at least they are commonly used for that).

-1 to get your attention, I don't really have any issues with this patch as is.

Patch-set: 1