Update patch set 2
Patch Set 2:

I15fbeec is 100% unrelated to what I do here. I'm documenting the status quo. As it is. Right now. The RDF export is already there and exporting exactly what I'm documenting here. This patch does not change the status quo.

It fixes the bug that the status quo is not documented.

Please upload your own patch and make it based on top of this, so this patch that does nothing but documenting the current status quo can be merged. Later, when the status quo changed, you can change the documentation together with that. This does not mean this patch must be blocked. Why should it? I really don't get it.

Patch-set: 2
Label: Verified=0