Fix GetClaimsStatementFilter not failing correctly with invalid rank
Luckily this is not critical in production, because the API takes care
that no invalid rank can make it that far. See GetClaims::getAllowedParams.

But when looking at the two classes I'm touching in this patch, there is
no safety. Both filters (by rank and by property) had bugs and failed with
type errors and "undefined variable" errors.

The StatementRankSerializer did not had any safety.

This also increases code coverage (which was the reason why I started
working on this).

Change-Id: I750c5c569cd336ff67f64744d23ab1666ce790b6